Profile after reg fix + reviews stars fix

+++ app/src/main/java/finalproj/dressapp/httpclient/models/UserModel.java
@@ -18,7 +18,7 @@ public class UserModel {
     public String lastName;
 
     @SerializedName("averageScore")
-    public int averageScore;
+    public float averageScore;
 
     @SerializedName("reviewQuantity")
     public int reviewQuantity;
